move ResolvedObjects
